+++ R/sector_aggregations.R
@@ -91,14 +91,6 @@ calculate_sector_agg_eta_fu <- function(.psut_data,
                                         useful = IEATools::all_stages$useful,
                                         eta_fu = Recca::efficiency_cols$eta_fu) {
 
-  # filtered_data <- .psut_data |>
-  #   dplyr::filter(.data[[country]] %in% countries, .data[[year]] %in% years)
-  # rm(.psut_data)
-  # gc()
-#
-#   if (nrow(filtered_data) == 0) {
-#     return(NULL)
-#   }
     if (nrow(.psut_data) == 0) {
       return(NULL)
     }
@@ -110,7 +102,7 @@ calculate_sector_agg_eta_fu <- function(.psut_data,
                                   notation = notation,
                                   pattern_type = pattern_type,
                                   prepositions = prepositions,
-                                  U = U, U_feed = U_feed, Y = Y,
+                                  U_eiou = U_eiou, Y = Y,
                                   by = "Sector",
                                   net_aggregate_demand = net_aggregate_demand,
                                   gross_aggregate_demand = gross_aggregate_demand) |>