+++ _targets.R
@@ -10,12 +10,13 @@ library(targets)
 
 # Set the countries to be analyzed.
 # countries <- c("GBR", "USA", "MEX")
+# countries <- c("ZWE", "USA", "WRLD")
 # countries <- "USA"
 # countries <- "WRLD"
 # countries <- "CHNM"
-countries <- c("GHA", "ZAF")
+# countries <- "GHA"
 # countries <- "all" # Run all countries in the PSUT target.
-# countries <- c(PFUDatabase::canonical_countries, "WRLD") |> as.character()
+countries <- c(PFUPipelineTools::canonical_countries, "WRLD") |> as.character()
 # Countries with unique allocation data plus BEL and TUR (for Pierre).
 # countries <- c("BRA", "CAN", "CHNM", "DEU", "DNK", "ESP", "FRA", "GBR", "GHA", "GRC",
 #                "HKG", "HND", "IDN", "IND", "JOR", "JPN", "KOR", "MEX", "NOR", "PRT",
@@ -23,8 +24,8 @@ countries <- c("GHA", "ZAF")
 
 
 # Set the years to be analyzed.
-# years <- 1960:2020
-years <- 1971:1973
+years <- 1960:2020
+# years <- 1971:1973
 # years <- 1971:1978
 # years <- 1971
 # years <- 1960:1961
@@ -36,7 +37,7 @@ do_chops <- FALSE
 # psut_release <- "20230309T184624Z-7ace5"  # v0.9 (USA only)
 # psut_release <- "20221109T152414Z-7d7ad"  # v1.0 (with matrix objects)
 # psut_release <- "20230312T211924Z-007da"  # v1.0 (with Matrix objects)
-psut_release <- "20230606T153820Z-d3dfe"    # v1.1 (with Matrix objects)
+psut_release <- "20230618T131003Z-4c70f"    # v1.1 (with Matrix objects)
 # psut_release <- "20221219T143657Z-964a6"  # For WRLD
 # psut_release <- "20230130T150642Z-631e2"  # For WRLD, 1971
 # psut_release <- "20230130T192359Z-1d3ec"  # For WRLD, 1971-2019
@@ -69,10 +70,6 @@ if (startsWith(sys_info[["nodename"]], "Mac")) {
 } else {
   stop("Unknown system in _targets.R for PFUAggDatabase. Can't set input and output locations.")
 }
-
-print(setup)
-stop()
-
 
 # Set up for multithreaded work on the local machine.
 future::plan(future.callr::callr)